+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+<a name="3.0.0"></a>
+# [3.0.0](https://github.com/atomastic/registry) (2021-02-20)
+* Move to PHP 7.4
+* add macroable functionality
+* use full power of Arrays Component by extending it.
+* remove final class
+* general code refactoring.
+* improve tests workflow.
+
 <a name="2.0.0"></a>
 # [2.0.0](https://github.com/atomastic/registry) (2020-12-02)
 * use new Arrays 2.0.0